+++ docs/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add sponsor info package
+- Add `b` tag global font colour
+
+### Changed
+
+- Update `p` tag font family to Hindˇ
+
 ## [1.5.1](https://github.com/hackmcgill/mchacks7/tree/1.5.1) - 2020-01-15
 
 ### Fixed
@@ -41,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactor `js` files to `jsx`
 
-### Fixed
+### Fixed``
 
 - Fix typo in `/privacy` page